+++ src/Rask.Server/RaskEndpointExtensions.cs
@@ -213,6 +213,7 @@ public static class RaskEndpointExtensions
         services.AddScoped<IPerformance, Performance>();
         services.AddScoped<IIndexedDb, IndexedDb>();
         services.AddScoped<IFileSystemAccess, FileSystemAccess>();
+        services.AddScoped<IWebAuthn, WebAuthn>();
         services.AddScoped<ICookies, Cookies>();
         services.AddScoped<IPermissions, Permissions>();
         services.AddScoped<IVibration, Vibration>();