+++ src/Rask.Wasm/WasmHostBuilder.cs
@@ -53,6 +53,7 @@ public sealed class WasmHostBuilder
         Services.AddSingleton<IPerformance, Performance>();
         Services.AddSingleton<IIndexedDb, IndexedDb>();
         Services.AddSingleton<IFileSystemAccess, FileSystemAccess>();
+        Services.AddSingleton<IWebAuthn, WebAuthn>();
         Services.AddSingleton<ICookies, Cookies>();
         Services.AddSingleton<IPermissions, Permissions>();
         Services.AddSingleton<IShare, Share>();